fix: change BigDecimal to Long in HoldingsEntity

+++ src/main/java/com/github/leo51645/assetflow/holdings/domain/entity/HoldingsEntity.java
@@ -30,7 +30,7 @@ public class HoldingsEntity {
 
     @Setter
     @Column(nullable = false)
-    private BigDecimal quantity;
+    private Long quantity;
 
     @Setter
     @Column(nullable = false)
